Issue # SC-1025: rest client impl complete with shifting esutil's method to tcp impl
